radiation wrapper class
