--- workflows/prognostic_c48_run/runtime/radiation_scheme/__init__.py
@@ -1,1 +1,0 @@
-from .config import RadiationConfig, init_radiation_driver
